+++ MoltenVK/MoltenVK/Commands/MVKCommandPipelineStateFactoryShaderSource.h
@@ -133,6 +133,24 @@ kernel void cmdClearColorImage2DInt(texture2d<int, access::write> dst [[ texture
     dst.write(clearValue, pos);                                                                                 \n\
 }                                                                                                               \n\
                                                                                                                 \n\
+kernel void cmdResolveColorImage2DFloat(texture2d<float, access::write> dst [[ texture(0) ]],                   \n\
+                                       texture2d_ms<float, access::read> src [[ texture(1) ]],                  \n\
+                                       uint2 pos [[thread_position_in_grid]]) {                                 \n\
+    dst.write(src.read(pos, 0), pos);                                                                           \n\
+}                                                                                                               \n\
+                                                                                                                \n\
+kernel void cmdResolveColorImage2DUInt(texture2d<uint, access::write> dst [[ texture(0) ]],                     \n\
+                                       texture2d_ms<uint, access::read> src [[ texture(1) ]],                   \n\
+                                       uint2 pos [[thread_position_in_grid]]) {                                 \n\
+    dst.write(src.read(pos, 0), pos);                                                                           \n\
+}                                                                                                               \n\
+                                                                                                                \n\
+kernel void cmdResolveColorImage2DInt(texture2d<int, access::write> dst [[ texture(0) ]],                       \n\
+                                      texture2d_ms<int, access::read> src [[ texture(1) ]],                     \n\
+                                      uint2 pos [[thread_position_in_grid]]) {                                  \n\
+    dst.write(src.read(pos, 0), pos);                                                                           \n\
+}                                                                                                               \n\
+                                                                                                                \n\
 typedef struct {                                                                                                \n\
     uint32_t srcRowStride;                                                                                      \n\
     uint32_t srcRowStrideHigh;                                                                                  \n\