+++ MoltenVK/MoltenVK/GPUObjects/MVKRenderPass.h
@@ -103,8 +103,8 @@ public:
 										 uint32_t passIdx,
 										 VkExtent2D framebufferExtent,
 										 uint32_t framebufferLayerCount,
-										 const MVKArrayRef<MVKImageView*>& attachments,
-										 const MVKArrayRef<VkClearValue>& clearValues,
+										 const MVKArrayRef<MVKImageView*> attachments,
+										 const MVKArrayRef<VkClearValue> clearValues,
 										 bool isRenderingEntireAttachment,
                                          bool loadOverride = false);
 
@@ -113,7 +113,7 @@ public:
 	 * when the render area is smaller than the full framebuffer size.
 	 */
 	void populateClearAttachments(MVKClearAttachments& clearAtts,
-								  const MVKArrayRef<VkClearValue>& clearValues);
+								  const MVKArrayRef<VkClearValue> clearValues);
 
 	/**
 	 * Populates the specified vector with VkClearRects for clearing views of a specified multiview
@@ -127,15 +127,16 @@ public:
 	/** If a render encoder is active, sets the store actions for all attachments to it. */
 	void encodeStoreActions(MVKCommandEncoder* cmdEncoder,
 							bool isRenderingEntireAttachment,
-							const MVKArrayRef<MVKImageView*>& attachments,
+							const MVKArrayRef<MVKImageView*> attachments,
 							bool storeOverride = false);
 
-	/** Constructs an instance for the specified parent renderpass. */
+	/** Resolves any resolve attachments that cannot be handled by native Metal subpass resolve behavior. */
+	void resolveUnresolvableAttachments(MVKCommandEncoder* cmdEncoder, const MVKArrayRef<MVKImageView*> attachments);
+
 	MVKRenderSubpass(MVKRenderPass* renderPass, const VkSubpassDescription* pCreateInfo,
 					 const VkRenderPassInputAttachmentAspectCreateInfo* pInputAspects,
 					 uint32_t viewMask);
 
-	/** Constructs an instance for the specified parent renderpass. */
 	MVKRenderSubpass(MVKRenderPass* renderPass, const VkSubpassDescription2* pCreateInfo);
 
 private:
@@ -187,7 +188,8 @@ public:
                                                    MVKRenderSubpass* subpass,
                                                    bool isRenderingEntireAttachment,
 												   bool isMemorylessAttachment,
-                                                   bool hasResolveAttachment,
+												   bool hasResolveAttachment,
+												   bool canResolveFormat,
                                                    bool isStencil,
                                                    bool loadOverride = false);
 
@@ -197,6 +199,7 @@ public:
 						   bool isRenderingEntireAttachment,
 						   bool isMemorylessAttachment,
 						   bool hasResolveAttachment,
+						   bool canResolveFormat,
 						   uint32_t caIdx,
 					   	   bool isStencil,
 						   bool storeOverride = false);
@@ -225,6 +228,7 @@ protected:
 									 bool isRenderingEntireAttachment,
 									 bool isMemorylessAttachment,
 									 bool hasResolveAttachment,
+									 bool canResolveFormat,
 									 bool isStencil,
 									 bool storeOverride);
 	void validateFormat();